% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{add_motif_db}
\alias{add_motif_db}
\title{add_motif_db}
\usage{
add_motif_db(object, motif.db, db.name)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{motif.db}{\code{PFMatrixList} or \code{PWMatrixList}}

\item{db.name}{the name of the transcription factor motif database}
}
\value{
a \code{CellSpace} object containing the motif embedding matrix, in the \code{motif.emb} slot, and the corresponding similarity Z-scores, in the \code{misc} slot
}
\description{
Computes the CellSpace embedding and activity scores of transcription factor motifs.
}
