% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\docType{class}
\name{CellSpace-class}
\alias{CellSpace-class}
\title{The CellSpace Class}
\description{
The CellSpace class stores CellSpace embedding and
related information needed for performing downstream analyses.

The CellSpace class stores CellSpace embedding and
related information needed for performing downstream analyses.
}
\section{Slots}{

\describe{
\item{\code{project}}{title of the project}

\item{\code{emb.file}}{the .tsv output of CellSpace containing the embedding matrix for cells and k-mers}

\item{\code{cell.emb}}{the embedding matrix for cells}

\item{\code{kmer.emb}}{the embedding matrix for k-mers}

\item{\code{motif.emb}}{the embedding matrix for TF motifs}

\item{\code{meta.data}}{data frame containing meta-information about each cell}

\item{\code{dim}}{the dimensions of the CellSpace embeddings}

\item{\code{k}}{the length of DNA k-mers}

\item{\code{similarity}}{the similarity function in hinge loss}

\item{\code{p}}{the embedding of an entity equals the sum of its M feature embedding vectors divided by M^p}

\item{\code{label}}{cell label prefix}

\item{\code{neighbors}}{list containing nearest neighbor graphs}

\item{\code{reductions}}{list containing dimensional reductions}

\item{\code{misc}}{list containing miscellaneous objects}

\item{\code{project}}{title of the project}

\item{\code{emb.file}}{the .tsv output of CellSpace containing the embedding matrix for cells and k-mers}

\item{\code{cell.emb}}{the embedding matrix for cells}

\item{\code{kmer.emb}}{the embedding matrix for k-mers}

\item{\code{motif.emb}}{the embedding matrix for TF motifs}

\item{\code{meta.data}}{data frame containing meta-information about each cell}

\item{\code{dim}}{the dimensions of the CellSpace embeddings}

\item{\code{k}}{the length of DNA k-mers}

\item{\code{similarity}}{the similarity function in hinge loss}

\item{\code{p}}{the embedding of an entity equals the sum of its M feature embedding vectors divided by M^p}

\item{\code{label}}{cell label prefix}

\item{\code{neighbors}}{list containing nearest neighbor graphs}

\item{\code{reductions}}{list containing dimensional reductions}

\item{\code{misc}}{list containing miscellaneous objects}
}}

