% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{CellSpace}
\alias{CellSpace}
\title{CellSpace}
\usage{
CellSpace(
  emb.file,
  cell.names = NULL,
  meta.data = NULL,
  project = NULL,
  similarity = "cosine",
  p = 0.5,
  label = "__label__"
)
}
\arguments{
\item{emb.file}{the .tsv output of CellSpace containing the embedding matrix for cells and k-mers}

\item{cell.names}{vector of unique cell names}

\item{meta.data}{a \code{data.frame} containing meta-information about each cell}

\item{project}{title of the project}

\item{similarity}{the similarity function in hinge loss}

\item{p}{the embedding of an entity equals the sum of its M feature embedding vectors divided by M^p}

\item{label}{cell label prefix}
}
\value{
a new \code{CellSpace} object
}
\description{
Generates an object from the CellSpace class.
}
