% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{cosine_similarity}
\usage{
cosine_similarity(x, y = NULL)
}
\arguments{
\item{x}{an embedding matrix}

\item{y}{an embedding matrix with compatible dimensions to \code{x}, or \code{NULL}, in which case \code{y=x}}
}
\value{
a matrix containing the cosine similarity between rows of \code{x} and \code{y}
}
\description{
Computes cosine similarity in the embedding space.
}
