% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{DNA_sequence_embedding}
\alias{DNA_sequence_embedding}
\title{DNA_sequence_embedding}
\usage{
DNA_sequence_embedding(object, seq)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{seq}{a DNA sequence}
}
\description{
Map a DNA sequence to the embedding space.
}
