% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{embedding_distance}
\alias{embedding_distance}
\title{embedding_distance}
\usage{
embedding_distance(x, y = NULL, distance = c("cosine", "angular"))
}
\arguments{
\item{x}{an embedding matrix}

\item{y}{an embedding matrix with compatible dimensions to \code{x}, or \code{NULL}, in which case \code{y=x}}

\item{distance}{the distance metric, either 'cosine' or 'angular', to compute from the cosine similarity}
}
\value{
a matrix containing the distance between rows of \code{x} and \code{y}, computed from their cosine similarity
}
\description{
Computes distance in the embedding space based on cosine similarity.
}
