% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{find_clusters}
\alias{find_clusters}
\title{find_clusters}
\usage{
find_clusters(object, graph = "cells_snn", ...)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{graph}{name of the nearest neighbor graph in the \code{neighbors} slot used to find clusters}

\item{...}{arguments passed to \code{Seurat::FindClusters}}
}
\value{
a \code{CellSpace} object with the cell clusters added to the \code{meta.data} slot
}
\description{
Find clusters in a nearest neighbor graph built from the CellSpace embedding.
}
