% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{find_neighbors}
\alias{find_neighbors}
\title{find_neighbors}
\usage{
find_neighbors(
  object,
  n.neighbors = 30,
  emb = object@cell.emb,
  emb.name = "cells",
  ...
)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{n.neighbors}{the number of nearest neighbors for the KNN algorithm}

\item{emb}{the embedding matrix used to create the nearest neighbor graphs}

\item{emb.name}{prefix for the graph names that will be added to the \code{neighbors} slot}

\item{...}{arguments passed to \code{Seurat::FindNeighbors}}
}
\value{
a \code{CellSpace} object containing nearest neighbor and shared nearest neighbor graphs in the \code{neighbors} slot
}
\description{
Builds a nearest neighbor graph and shared nearest neighbor graph from the CellSpace embedding.
}
