% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\name{merge_small_clusters}
\alias{merge_small_clusters}
\title{merge_small_clusters}
\usage{
merge_small_clusters(
  object,
  clusters,
  min.cells = 10,
  graph = "cells_snn",
  seed = 1
)

merge_small_clusters(
  object,
  clusters,
  min.cells = 10,
  graph = "cells_snn",
  seed = 1
)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{clusters}{a vector of cluster labels or the name of a column, in the \code{meta.data} slot, containing cluster labels}

\item{min.cells}{any cluster with fewer cells than \code{min.cells} will be merged with the nearest cluster}

\item{graph}{a nearest neighbor graph, or the name of a nearest neighbor graph in the \code{neighbors} slot, used to find clusters}
}
\value{
new cluster labels

new cluster labels
}
\description{
Merge cells from small clusters with the nearest clusters.

Merge cells from small clusters with the nearest clusters.
}
