% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\name{run_UMAP}
\alias{run_UMAP}
\title{run_UMAP}
\usage{
run_UMAP(object, emb = object@cell.emb, graph = NULL, name = "cells_UMAP", ...)

run_UMAP(object, emb = object@cell.emb, graph = NULL, name = "cells_UMAP", ...)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{emb}{the embedding matrix used to compute the UMAP embedding}

\item{graph}{name of the nearest neighbor graph in the \code{neighbors} slot used to compute the UMAP embedding, used only if \code{emb} is NULL}

\item{name}{name of the lower-dimensional embedding that will be added to the \code{reductions} slot}

\item{...}{arguments passed to \code{Seurat::RunUMAP}}
}
\value{
a \code{CellSpace} object containing a UMAP embedding in the \code{reductions} slot

a \code{CellSpace} object containing a UMAP embedding in the \code{reductions} slot
}
\description{
Compute a UMAP embedding from the CellSpace embedding.

Compute a UMAP embedding from the CellSpace embedding.
}
