% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{cosine_similarity}
\usage{
cosine_similarity(x, y = NULL)

cosine_similarity(x, y = NULL)
}
\arguments{
\item{x}{an embedding matrix}

\item{y}{NULL, in which case \code{y=x}, or an embedding matrix with compatible dimensions to \code{x}}
}
\description{
Cosine similarity in the embedding space.

Cosine similarity in the embedding space.
}
