% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\name{embedding_distance}
\alias{embedding_distance}
\title{embedding_distance}
\usage{
embedding_distance(x, y = NULL, distance = c("cosine", "angular"))

embedding_distance(x, y = NULL, distance = c("cosine", "angular"))
}
\arguments{
\item{x}{an embedding matrix}

\item{y}{NULL, in which case \code{y=x}, or an embedding matrix with compatible dimensions to \code{x}}

\item{distance}{the distance metric, either 'cosine' or 'angular', to compute from the cosine similarity}
}
\description{
Distance in the embedding space based on cosine similarity.

Distance in the embedding space based on cosine similarity.
}
