% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace-temp.R, R/CellSpace.R
\name{motif_embedding}
\alias{motif_embedding}
\title{motif_embedding}
\usage{
motif_embedding(object, PWM)

motif_embedding(object, PWM)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{PWM}{\code{PFMatrix} or \code{PWMatrix}}

\item{method}{???}

\item{min.prob}{???}

\item{convert.PWM}{???}
}
\description{
Map a motif to the embedding space.

Map a motif to the embedding space.
}
