% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{motif_embedding}
\alias{motif_embedding}
\title{motif_embedding}
\usage{
motif_embedding(object, PWM)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{PWM}{position weight matrix or position frequency matrix}
}
\description{
Map a motif to the embedding space.
}
