% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSpace.R
\name{run_UMAP}
\alias{run_UMAP}
\title{run_UMAP}
\usage{
run_UMAP(
  object,
  n.neighbors = 30,
  emb = object@cell.emb,
  graph = NULL,
  name = "cells_UMAP",
  ...
)
}
\arguments{
\item{object}{a \code{CellSpace} object}

\item{n.neighbors}{the number of nearest neighbors for UMAP}

\item{emb}{the embedding matrix used to compute the UMAP embedding}

\item{graph}{name of the nearest neighbor graph in the \code{neighbors} slot used to compute the UMAP embedding, used only if \code{emb} is NULL}

\item{name}{name of the lower-dimensional embedding that will be added to the \code{reductions} slot}

\item{...}{arguments passed to \code{Seurat::RunUMAP}}
}
\value{
a \code{CellSpace} object containing a UMAP embedding in the \code{reductions} slot
}
\description{
Compute a UMAP embedding from the CellSpace embedding.
}
